Refactor OpenWire SSL Tests and Enhance SSL Configuration
- Updated OpenwireSslSlowListenerTest to improve namespace formatting.
- Refactored OpenwireSslTempDestinationTest to replace inheritance from TempDestinationTest with CMSTestFixture, adding new test cases for message handling and connection behavior.
- Enhanced OpenwireSslTransactionTest and OpenwireSslXATransactionsTest to utilize Google Test framework, improving structure and readability.
- Updated IntegrationCommon to include SSL peer verification details.
- Modified CMake presets to enable SSL support.
- Adjusted docker-compose and SSL configuration files to reflect new paths for ActiveMQ and certificate files.

+++ activemq-cpp/src/main/decaf/internal/net/SocketFileDescriptor.cpp
@@ -23,7 +23,7 @@ using namespace decaf::internal;
 using namespace decaf::internal::net;
 
 ////////////////////////////////////////////////////////////////////////////////
-SocketFileDescriptor::SocketFileDescriptor( long value ) : FileDescriptor( value, false ) {
+SocketFileDescriptor::SocketFileDescriptor( intptr_t value ) : FileDescriptor( value, false ) {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -31,6 +31,6 @@ SocketFileDescriptor::~SocketFileDescriptor() {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-long SocketFileDescriptor::getValue() const {
+intptr_t SocketFileDescriptor::getValue() const {
     return this->descriptor;
 }
